Moved the lbann_library files into the utils directory.  Cleaned up a string print statement.
