+++ src/CMakeLists.txt
@@ -3,7 +3,6 @@ set_full_path(THIS_DIR_SOURCES
   Elemental_extensions.cpp
   base.cpp
   comm.cpp
-  lbann_library.cpp
   )
 
 # Add the subdirectories
